Closes #188: Move tests for components into their corresponding component folders
