+++ src/react-components/package.json
@@ -16,8 +16,8 @@
         "build": "tsc",
         "dist": "rm -rf dist && webpack --mode production --progress --display-modules --config webpack.prod.js",
         "watch-webpack": "webpack --mode development --watch --progress --display-modules --config webpack.prod.js",
-        "lint": "tslint -c tslint.json src/__tests__/components/*.tsx src/components/**/**/*.tsx src/components/**/**/**/*.tsx src/utils/*.tsx",
-        "test": "rm -rf lib && npm run lint && tsc && mocha --require lib/__tests__/setupEnzyme.js --recursive lib/__tests__"
+        "lint": "tslint -c tslint.json src/__tests__/helpers/*.tsx src/components/**/**/*.tsx src/components/**/**/**/*.tsx src/utils/*.tsx",
+        "test": "rm -rf lib && npm run lint && tsc && mocha --require lib/__tests__/setupEnzyme.js --recursive lib/__tests__ --recursive lib/components"
     },
     "keywords": [],
     "devDependencies": {
