+++ src/react-components/src/components/Button/Button-test.tsx
@@ -4,7 +4,7 @@ import * as Enzyme from "enzyme";
 import * as React from "react";
 import * as Mocha from "mocha";
 
-import Button from "../../components/Button/Button";
+import Button from "./Button";
 
 describe("Button", () => {
     let wrapper: Enzyme.ShallowWrapper<{}, {}>;
