+++ src/react-components/src/components/Heading/Heading-test.tsx
@@ -3,7 +3,7 @@ import * as Enzyme from "enzyme";
 import * as React from "react";
 import * as Mocha from "mocha";
 
-import Heading from "../../components/Heading/Heading";
+import Heading from "./Heading";
 
 describe("Section Headings", () => {
     let wrapper: Enzyme.ShallowWrapper<{}, {}>;
