+++ src/react-components/src/components/HelperErrorText/HelperErrorText-test.tsx
@@ -3,7 +3,7 @@ import * as Enzyme from "enzyme";
 import * as React from "react";
 import * as Mocha from "mocha";
 
-import HelperErrorText from "../../components/HelperErrorText/HelperErrorText";
+import HelperErrorText from "./HelperErrorText";
 
 describe("HelperErrorText Test", () => {
     it("Renders HelperErrorText", () => {
