+++ src/react-components/src/components/Image/Image-test.tsx
@@ -3,7 +3,7 @@ import * as Enzyme from "enzyme";
 import * as React from "react";
 import * as Mocha from "mocha";
 
-import Image from "../../components/Image/Image";
+import Image from "./Image";
 
 describe("Images", () => {
     let wrapper: Enzyme.ShallowWrapper<{}, {}>;
