+++ src/react-components/src/components/UnorderedList/UnorderedList-test.tsx
@@ -3,7 +3,7 @@ import { expect } from "chai";
 import * as Enzyme from "enzyme";
 import * as React from "react";
 
-import UnorderedList from "../../components/UnorderedList/UnorderedList";
+import UnorderedList from "./UnorderedList";
 
 describe("Link With Underline", () => {
     let wrapper: Enzyme.ReactWrapper<{}, {}>;
